Fix bug that was causing hector exceptions to crash MC runs
The issue here was that when we try to turn the error condition into a
message, by writing `message(e)`, where `e` was the error condition
object, foreach detects this as an error and aborts.  Other looping
constructs don't have this problem, which is probably why we haven't
seen it before now.

Anyhow, the fix is to call `conditionMessage` on `e`, which gives us
the text that we want, while stripping it of its error context.