Update PCA function tests to use new principal components.
We also had to extend the number of components that we keep for these
tests.  That's because one of our tests is actually a little bogus.
We test that the project_climate and reconstruct_climate functions are
inverses of one another, but that's technically only true if we keep
_all_ of the components.  When you trim the components down, like we
do for the test, you're no longer guaranteed that the two operations
are inverses.  Still, it's close enough to verify that the functions
are working as intended.