Add comparison data sets for likelihood functions using principal components.
The concentration-driven version of this includes a heat flux value
that can be used to further constrain the diffusivity parameter.  This
value is absent from the emissions-driven version because the data
needed to create it is not available for most of the models that
provided data for the emissions-driven runs.